Reorganize Mock modules
