+++ lib/davinci_dtr_test_kit/dtr_smart_app_suite.rb
@@ -5,12 +5,12 @@ require_relative 'auth_groups/oauth2_authentication_group'
 require_relative 'client_groups/resp_assist_device/dtr_smart_app_questionnaire_workflow_group'
 require_relative 'client_groups/dinner_static/dtr_smart_app_questionnaire_workflow_group'
 require_relative 'client_groups/dinner_adaptive/dtr_smart_app_questionnaire_workflow_group'
-require_relative 'endpoints/mock_authorization/mock_authorization'
+require_relative 'endpoints/mock_authorization'
 require_relative 'endpoints/mock_authorization/authorize_endpoint'
 require_relative 'endpoints/mock_authorization/token_endpoint'
 require_relative 'endpoints/mock_payer/questionnaire_package_endpoint'
 require_relative 'endpoints/mock_payer/next_question_endpoint'
-require_relative 'endpoints/mock_ehr/mock_ehr'
+require_relative 'endpoints/mock_ehr'
 require_relative 'endpoints/mock_ehr/questionnaire_response_endpoint'
 require_relative 'endpoints/mock_ehr/fhir_get_endpoint'
 require_relative 'mock_auth_server'
