+++ lib/davinci_dtr_test_kit/endpoints/mock_ehr/fhir_get_endpoint.rb
@@ -1,5 +1,5 @@
-require_relative '../mock_authorization/mock_authorization'
-require_relative 'mock_ehr'
+require_relative '../mock_authorization'
+require_relative '../mock_ehr'
 
 module DaVinciDTRTestKit
   module MockEHR
